Graph model object working properly for outliers and normal graph display
Deleted other models as they shouldn't have been models in the first place. The graph object now gets created everytime the page is submitted and the values are saved into the database to be tracked over time. Still need to do something with post/redirect/get so that a refresh doesnt submit the form again.